requirements.txt
opencv-python-headless
PyYAML
fastapi
uvicorn
SQLAlchemy
python-dotenv
